Add flag to trust rev-proxy header for TLS.
When behind a reverse-proxy that is handling TLS, we need to generate the
correct references and links so that resources can be included and there are no
links back to the http site. To accomplish this, we add a flag to the tls config
that says to trust the X-Forwarded-Proto header and doesn't require a key or
cert.